Bank Acount Project in python
The Bank Account Project is a simple Python command-line application that simulates basic banking operations. It includes an Account class to manage individual accounts, an ATM class to interact with these accounts, and a main application that lets users create new accounts, log in to existing ones, and perform actions like checking balance, withdrawing, and depositing funds. User data is stored in a text file for persistence. The project provides a basic banking experience, showcasing object-oriented programming and file handling.